Optimised solver
Tries to solve inbetween steps now to see if it needs more calibratrion
to be solved. Also makes it only attempt to solve when it can be an
optimal solution (the solution is only optimal if the order of it goes
down from H to A). also removed small workaround