Quest tab filter (#2)
* Add quest filtering plugin

This plugin allows for users to filter their quest list by type (f2p, members, & miniquests), as well as by progress status (completed, in progress, and not started). Any combination of the 6 filtering options is valid.

The filtering is done during the initialization of the quest list widget (script 1350.rs2asm AKA QuestListInit.rs2asm) and during the initialization of each quest list section (script 1351.rs2asm AKA QuestListItemInit.rs2asm). A side effect of this is that the quest list must be reinitialized in order for configuration changes to be applied. This can be done simply by switching to the achievement diary list and back to the quest tab list. Whenever a quest list filter is changed through the configuration menu, a text panel will appear and give reinitialization instructions to the user. Once the list is reinitialized, the text panel will disappear and the selected filters will have been applied.

* Make filtering instant and remove dependency on custom 1350 script

These changes will make all filtering of the quest list instantaneous. Any time a filtering config is changed, a call is made to script 1350 to reinitialize the quest list. This call is made from within the QuestTabPlugin class.

All logic from the custom 1350 script - QuestListInit.rs2asm - has been moved to the custom 1351 script - QuestListItemInit.rs2asm. Since 1350 is updated every time a quest is added and 1351 is not, removing dependency on the custom 1350 script will greatly reduce the amount of maintenance required to keep the plugin working.